new tables for seats

+++ src/components/cart/cart.js
@@ -1,33 +1,24 @@
-import React, { Fragment, useEffect } from 'react';
+import React, { Fragment } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
-import { cartLoad } from '../../redux/actions/cartAction';
 import './cart.css'
 import { deleteAll, deleteFlightReservation, buyTicket } from '../../redux/actions/cartAction'
 import Button from '@material-ui/core/Button';
 import { useHistory } from "react-router-dom";
 import Swal from 'sweetalert2'
-import TableCreator from '../tableCreator/tableCreator'
+import MaterialTable from 'material-table'
 
 export default function Cart() {
   const dispatch = useDispatch();
-  const cart  = useSelector(state => state.cartReducer.flights);
+  const cart = useSelector(state => state.cartReducer.flights);
   const login = useSelector(store => store.login);
   let history = useHistory();
-  useEffect(() => {
-    getAllCart()
-  // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [])
+
 
   const onFlightsClick = e => {
     history.push("/");
   }
 
-  const getAllCart = () => {
-    dispatch(cartLoad(login.id))
-  }
-
-  const deleteFlight = async (flightSelect) => {
-
+  const deleteFlight = async (e, flightSelect) => {
     const { value: deleteAlert } = await Swal.fire({
       title: 'Estas seguro?',
       text: "Desea eliminar este pasaje?",
@@ -38,14 +29,28 @@ export default function Cart() {
       confirmButtonText: 'SI',
       cancelButtonText: 'NO'
     })
-    
+
     if (deleteAlert) {
       dispatch(deleteFlightReservation(flightSelect, login.id))
     }
   }
 
-  const deleteAllflights = () => {
-    dispatch(deleteAll(login.id))
+  const deleteAllflights = async () => {
+    const { value: deleteAlert } = await Swal.fire({
+      title: 'Estas seguro?',
+      text: "Desea eliminar todos los tickets?",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'SI',
+      cancelButtonText: 'NO'
+    })
+
+    if (deleteAlert) {
+      dispatch(deleteAll(login.id))
+    }
+
   }
 
   const buyTicketsFlights = async (e) => {
@@ -67,27 +72,53 @@ export default function Cart() {
       return alert
     }
   }
-  const isCartEmpty = () => {
-    return cart.numberOfTickets !== 0
-  }
-  const columnName = [
-    { name: 'Origen' }, { name: 'Destino' }, { name: 'Salida' }, { name: 'Aerolinea' }, { name: 'Asiento' }, { name: 'Clase' }, { name: 'Desde' }, { name: 'Accion' }
-  ]
 
   return (
     <Fragment>
-      <TableCreator data={cart.tickets} columnName={columnName} titleButton="Elimina" buttonAction={deleteFlight} />
+      <MaterialTable
+        title="Asientos"
+        columns={[
+          { title: "Origen", field: "from" },
+          { title: "Destino", field: "to" },
+          { title: "Salida", field: "departure" },
+          { title: "Aerolinea", field: "airline" },
+          { title: "Asiento", field: "seatNumber" },
+          { title: "Clase", field: "seatType" },
+          { title: "Precio", field: "cost" },
+        ]}
+        data={cart.tickets}
+        options={
+          {
+            search: false,
+            paging: false,
+            actionsColumnIndex: -1,
+          }
+        }
+        actions={[
+          {
+            icon: "delete",
+            tooltip: 'Eliminar ticket',
+            onClick: deleteFlight
+          }
+        ]}
+        localization={
+          {
+            body: { emptyDataSourceMessage: "No hay items en el carrito" },
+            header: { actions: "Acciones" }
+          }
+        }
+      />
 
       <form onSubmit={buyTicketsFlights}>
         <div className="botones">
-          <Button variant="contained" color="primary" disabled={isCartEmpty} onClick={() => deleteAllflights()}>Limpiar carro</Button>
+          <Button variant="contained" color="primary" disabled={cart.numberOfTickets === 0} onClick={() => deleteAllflights()}>Limpiar carro</Button>
         </div>
         <h3 align="left">Total en el carrito: ${cart.totalCost}</h3>
         <div className="botonesInferior">
           <div className="botonVolver">
             <Button variant="contained" color="secondary" className="buttonVolver" onClick={onFlightsClick}>Volver</Button>
           </div>
-          <Button type="submit" variant="contained" color="primary" disabled={isCartEmpty}>Comprar</Button>
+          <Button type="submit" variant="contained" disabled={cart.numberOfTickets === 0} color="primary">Comprar</Button>
         </div>
       </form>
